Se realizó documentación de Mayorista, se hizo cambio de las ordenes de los minoristas no tienen tiempo de espera, se hizo cambio de tipo de variables de double a int.

+++ src/inventario/controlador/CMayorista.java
@@ -32,7 +32,7 @@ public class CMayorista {
         ventana.setVisible(true);
     }
     
-    public static void iniciarServidor(double p, double q,  int puerto) {
+    public static void iniciarServidor(int p, int q,  int puerto) {
         mayorista = new Mayorista(p, q, generador);
         try {
             mayorista.iniciar(puerto);
